backend/requirements.txt
Flask==2.2.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.3.1
Flask-PyMongo==2.3.0
Flask-Cors==3.0.10
pymongo[srv]==4.0.1
python-dotenv==0.19.2
dnspython==2.2.0
Werkzeug==2.2.2

